% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounting.R
\name{accountingmultiplier}
\alias{accountingmultiplier}
\title{Accounting Multiplier}
\usage{
accountingmultiplier(Af)
}
\arguments{
\item{Af}{prospensity matrix}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to generate multiplier
}
\author{
Tiara Dewi
}

