% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageprospensitymatrix.R
\name{adjustedmatrix}
\alias{adjustedmatrix}
\title{Average Prospensity Matrix}
\usage{
adjustedmatrix(newfsam, baris1, kolom1)
}
\arguments{
\item{newfsam}{data that contain endogen accounts only from getendogen.R}

\item{baris1}{number containing start row generate from user input}

\item{kolom1}{number containing start column generate from user input}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to generate the average prospensity matrix from data
}
\author{
Tiara Dewi
}

