% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{dekomposisi}
\alias{dekomposisi}
\title{Decomposition Result}
\usage{
dekomposisi(I, Ca, Oa, Ta)
}
\arguments{
\item{I}{injection matrix generate by getMatriksInjeksi function in injeksi.R}

\item{Ca}{Closed Loop matrix generate by getCa function in ClosedLoop.R}

\item{Oa}{Open Loop matrix generate by getOa function in OpenLoop.R}

\item{Ta}{Transfer Loop matrix generate by getTa function in Transfer.R}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to calculate the result of accounting multipier decomposition multiply by injection matrix
}
\author{
Tiara Dewi
}

