% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayresult.R
\name{displayResult}
\alias{displayResult}
\title{Display Result}
\usage{
displayResult(Ma, judul, kode)
}
\description{
support function to GUI Multiplier function. Not intended to be called directly. Visible after user had choose either function of Accounting Multiplier SAM, Accounting Multiplier Decomposition SAM, Fixed Price Multiplier SAM or Accounting Multiplier FSAM.
}
\author{
Tiara Dewi
}

