% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedprice.R
\name{fixedprice}
\alias{fixedprice}
\title{Fixed Price Multiplier}
\usage{
fixedprice(A, baris_start, faktor_prod, institusi, keg_prod, EXP, ei)
}
\arguments{
\item{A}{average prospensity matrix generate by adjustedmatrix function in averagprospensitymatrix.R}

\item{baris_start}{number containing start row generate from user input}

\item{faktor_prod}{number containing end row of production factors generate from user input}

\item{institusi}{number containing end row of institutions generate from user input}

\item{keg_prod}{number containing end row of economic activities generate from user input}

\item{EXP}{list containing row numbers of exogen accounts from user input}

\item{ei}{elasticiy matrix that user has input in getFixedprice function in getfixedprice.R}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly. Visible after user input elasticity matrix.
}
\details{
this function is to calculate matrix C (marginal prospensity matrix)
}
\author{
Tiara Dewi
}

