% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abintang.R
\name{getABintang}
\alias{getABintang}
\title{A*}
\usage{
getABintang(baris_start, faktor_prod, institusi, EXP, keg_prod, A, Ma1)
}
\arguments{
\item{baris_start}{number containing start row generate from user input}

\item{faktor_prod}{number containing end row of production factors generate from user input}

\item{institusi}{number containing end row of institutions generate from user input}

\item{EXP}{list containing row numbers of exogen accounts from user input}

\item{keg_prod}{number containing end row of economic activities generate from user input}

\item{A}{average prospensity matrix generate by adjustedmatrix function in averagprospensitymatrix.R}

\item{Ma1}{Ma1 matrix generate by getMa1 function in Ma1.R}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to calculate A* matrix for accounting multipier decomposition.
}
\author{
Tiara Dewi
}

