% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abintang2.R
\name{getABintang2}
\alias{getABintang2}
\title{A**}
\usage{
getABintang2(Abintang)
}
\arguments{
\item{Abintang}{A* matrix that generate by ABintang function in Abintang.R}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to calculate Abintang2 matrix for accounting multipier decomposition.
}
\author{
Tiara Dewi
}

