% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abintang3.R
\name{getABintang3}
\alias{getABintang3}
\title{A***}
\usage{
getABintang3(Abintang, Abintang2)
}
\arguments{
\item{Abintang}{A* matrix that generate by ABintang function in Abintang.R}

\item{ABintang2}{A** matrix that generate by ABintang2 function in Abintang2.R}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to calculate Abintang2 matrix for accounting multipier decomposition.
}
\author{
Tiara Dewi
}

