% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClosedLoop.R
\name{getCa}
\alias{getCa}
\title{Closed Loop}
\usage{
getCa(Ma1, Ma2, Ma3, I)
}
\arguments{
\item{Ma1}{Ma1 matrix that generate by getMa1 function in Ma1.R}

\item{Ma2}{Ma2 matrix that generate by getMa2 function in Ma2.R}

\item{Ma3}{Ma3 matrix that generate by getMa3 function in Ma3.R}

\item{I}{identity matrix}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to calculate Ca matrix for accounting multipier decomposition.
}
\author{
Tiara Dewi
}

