% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getendogen.R
\name{getEndogen}
\alias{getEndogen}
\title{Endogen Accounts}
\usage{
getEndogen(fsam, cinta, baris1, kolom1)
}
\arguments{
\item{fsam}{data}

\item{cinta}{list containing row numbers of exogen accounts from user input}

\item{baris1}{number containing start row generate from user input}

\item{kolom1}{number containing start column generate from user input}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to exclude exogen accounts from data
}
\author{
Tiara Dewi
}

