% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfixedprice.R
\name{getFixedprice}
\alias{getFixedprice}
\title{Input User for Fixed Price Multiplier SAM}
\usage{
getFixedprice()
}
\description{
support function to GUI Multiplier function. Not intended to be called directly. Visible beacuse user choose Accounting Multiplier in FSAM at GUI Multiplier
}
\author{
Tiara Dewi
}

