% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDecomposition.R
\name{getinjeksi}
\alias{getinjeksi}
\title{Input User for Accounting Multiplier Decomposition of SAM}
\usage{
getinjeksi()
}
\description{
support function to GUI Multiplier function. Not intended to be called directly. Visible because user choose Accounting Multiplier Decomposition in SAM at GUI Multiplier
}
\author{
Tiara Dewi
}

