% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ma1.R
\name{getMa1}
\alias{getMa1}
\title{Ma1}
\usage{
getMa1(baris_start, kolom_start, faktor_prod, institusi, EXP, keg_prod, A)
}
\arguments{
\item{baris_start}{number containing start row generate from user input}

\item{kolom_start}{number containing start column generate from user input}

\item{faktor_prod}{number containing end row of production factors generate from user input}

\item{institusi}{number containing end row of institutions generate from user input}

\item{EXP}{list containing row numbers of exogen accounts from user input}

\item{keg_prod}{number containing end row of economic activities generate from user input}

\item{A}{average prospensity matrix generate by adjustedmatrix function in averagprospensitymatrix.R}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to calculate Ma1 matrix for accounting multipier decomposition.
}
\author{
Tiara Dewi
}

