% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ma2.R
\name{getMa2}
\alias{getMa2}
\title{Ma2}
\usage{
getMa2(I, Abintang3)
}
\arguments{
\item{I}{identity matrix}

\item{ABintang3}{A*** matrix that generate by ABintang3 function in Abintang3.R}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to calculate Ma2 matrix for accounting multipier decomposition.
}
\author{
Tiara Dewi
}

