% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injeksi.R
\name{getMatriksInjeksi}
\alias{getMatriksInjeksi}
\title{Injection Matrix}
\usage{
getMatriksInjeksi(injeksi, baris, EXP, baris_start, A)
}
\arguments{
\item{injeksi}{value that will be injected in specified row}

\item{baris}{number containing row number that will be injected generate from user input}

\item{EXP}{list containing row numbers of exogen accounts from user input}

\item{baris_start}{number containing start row generate from user input}

\item{A}{average prospensity matrix generate by adjustedmatrix function in averagprospensitymatrix.R}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to calculate injection matrix
}
\author{
Tiara Dewi
}

