% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenLoop.R
\name{getOa}
\alias{getOa}
\title{Open Loop}
\usage{
getOa(Ma1, Ma2, I)
}
\arguments{
\item{Ma1}{Ma1 matrix that generate by getMa1 function in Ma1.R}

\item{Ma2}{Ma2 matrix that generate by getMa2 function on Ma2.R}

\item{I}{identity matrix}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to calculate Oa matrix for accounting multipier decomposition.
}
\author{
Tiara Dewi
}

