% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{getposisi}
\alias{getposisi}
\title{Input User for Accounting Multiplier SAM}
\usage{
getposisi()
}
\description{
support function to GUI Multiplier function. Not intended to be called directly. Visible because user choose Accounting Multiplier in SAM at GUI Multiplier
}
\author{
Tiara Dewi
}

