% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsam.R
\name{getposisifsam}
\alias{getposisifsam}
\title{Input User for Accounting Multiplier FSAM}
\usage{
getposisifsam()
}
\description{
support function to GUI Multiplier function. Not intended to be called directly. Visible beacuse user choose Accounting Multiplier in FSAM at GUI Multiplier
}
\author{
Tiara Dewi
}

