% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getsam}
\alias{getsam}
\title{Open File}
\usage{
getsam()
}
\description{
support function to GUI Multiplier function. Not intended to be called directly. Visible because user choose Open File in FSAM at GUI Multiplier.
}
\details{
the data must be in Microsoft Excel form ".xlsx" and saved by name sam in global environment
}
\author{
Tiara Dewi
}

