% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transfer.R
\name{getTa}
\alias{getTa}
\title{Transfer Loop}
\usage{
getTa(Ma1, I)
}
\arguments{
\item{Ma1}{Ma1 matrix that generate by getMa1 function in Ma1.R}

\item{I}{identity matrix}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to calculate Ta matrix for accounting multipier decomposition.
}
\author{
Tiara Dewi
}

