% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{GUI Multiplier}
\usage{
gui()
}
\description{
this function to create GUI that call all the functions in this package
}
\details{
this package was build to calculate the multiplier of Social Accounting Matrix (SAM) and Financial Social Accounting Matrix (FSAM)
}
\author{
Tiara Dewi
}

