% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ma3.R
\name{getMa3}
\alias{getMa3}
\title{Ma3}
\usage{
getMa3(I, Abintang, Abintang2)
}
\arguments{
\item{I}{identity matrix}

\item{Abintang}{A* matrix that generate by ABintang function in Abintang.R}

\item{Abintang2}{A** matrix that generate by ABintang2 function in Abintang2.R}
}
\description{
support function to GUI Multiplier function. Not intended to be called directly.
}
\details{
this function is to calculate Ma3 matrix for accounting multipier decomposition.
}
\author{
Tiara Dewi
}

